==> analysis/2-network.Rmd <==
---
title: "2-network"
author: "bernard-liew"
date: "2020-10-08"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Load package

```{r, include = FALSE}
knitr::opts_chunk$set (eval = FALSE)
```

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse,
               qgraph,
               stats,
               bootnet,
               igraph,
               mgm,
               NetworkComparisonTest,
               rio,
               furrr,
               cowplot,
               doParallel,
               huge,
               EGAnet)
```

## Import data

```{r}
rm(list = ls())
df <- readRDS("output/dat_odi_nest.RDS")

```

## Reorganize data

```{r}
temp1 <- df$raw %>% select (-data_mgm)
temp2 <- df$com %>% select (-data_mgm) %>% rename (data_com = data)
temp3 <- temp1 %>%
  inner_join(temp2, by = "time")

temp4 <- df$raw %>% select (-data) %>% rename (data = data_mgm)
temp5 <- df$com %>% select (-data) %>% rename (data_com = data_mgm)
temp6 <- temp4 %>%
  inner_join(temp5, by = "time")

df <- list (odi = temp3, 
            odi_mgm = temp6)

rm (list = ls(pattern = "temp"))
```


## Network analysis

### Set boot number

```{r}
B <- 2000
```


```{r}

stats_type <- c("edge", "strength", "betweenness", "expectedInfluence", "closeness")

my_huge <- function (df) {

  df[, grepl ("Q", names(df))] <- huge::huge.npn (df[, grepl ("Q", names(df))] )

  return (df)

}

```


### Analyse with groups as variable


```{r}
#### Analyse data with missing
dat <- df$odi_mgm 
set.seed(1)

res_mgm_raw <- dat %>%
  select (time, data) %>% 
  # remove cases with missing columns
  mutate (data = map (data, na.omit)) %>%
  # Transform data
  mutate (data = map(data, my_huge)) %>%
  # Estimate network
  mutate (nw = map (data,
                    estimateNetwork,
                    default = "mgm",
                     type= c("c", rep("g", 10)),
                     level= c(2, rep(1, 10)),
                     criterion  = "CV",            # we used cross validation to select optimal tuning parameter
                     nFolds = 10,            # using 10 folds
                     order = 2,                       # we only include second order interactions
                     binarySign = TRUE,
                     scale = TRUE,
                    .pbar = FALSE,
                    .signInfo = FALSE)) %>%
  # Get centrality measures
  mutate (centr = map (nw, centralityTable)) 


# Bootstrap by case to get stability of centrality
B <- 2000
plan (multisession)

res_mgm_raw <- res_mgm_raw %>%
  mutate (centr_stb = future_map (nw, 
                                  bootnet, 
                                  nBoots = B,
                                  type = "case",
                                  statistics = stats_type,
                                  .options = furrr_options(seed = TRUE)))

# Stability measure
res_mgm_raw <- res_mgm_raw %>%
  mutate (cor_stb = map (centr_stb,
                         corStability))

# Bootstrap by case to get stability of centrality
plan (multisession)

res_mgm_raw <- res_mgm_raw %>%
  mutate (edgewts = future_map (nw, 
                                  bootnet, 
                                  nBoots = B,
                                .options = furrr_options(seed = TRUE)))


saveRDS(res_mgm_raw, "output/mgm_raw.RDS")

```




```{r, eval = FALSE, include = FALSE}
#### Analyse data with complete
set.seed(1)

res_mgm_com <- dat %>%
  select (time, data_com) %>%
  rename (data = data_com) %>%
  # Transform data
  mutate (data = map(data, my_huge)) %>%
  # Estimate network
  mutate (nw = map (data,
                    estimateNetwork,
                    default = "mgm",
                     type= c("c", rep("g", 10)),
                     level= c(2, rep(1, 10)),
                     criterion  = "CV",            # we used cross validation to select optimal tuning parameter
                     nFolds = 10,            # using 10 folds
                     order = 2,                       # we only include second order interactions
                     binarySign = TRUE,
                     scale = TRUE,
                    .pbar = FALSE,
                    .signInfo = FALSE)) %>%
  # Get centrality measures
  mutate (centr = map (nw, centralityTable)) 


# Bootstrap by case to get stability of centrality
B <- 2000
plan (multisession)

res_mgm_com  <- res_mgm_com  %>%
  mutate (centr_stb = future_map (nw, 
                                  bootnet, 
                                  nBoots = B,
                                  type = "case",
                                  statistics = stats_type,
                                  .options = furrr_options(seed = TRUE)))

# Stability measure
res_mgm_com <- res_mgm_com  %>%
  mutate (cor_stb = map (centr_stb,
                         corStability))

# Bootstrap by case to get stability of centrality
plan (multisession)

res_mgm_com  <- res_mgm_com %>%
  mutate (edgewts = future_map (nw, 
                                  bootnet, 
                                  nBoots = B,
                                .options = furrr_options(seed = TRUE)))

saveRDS(res_mgm_com, "output/mgm_com.RDS")

```
