==> analysis/3-report.Rmd <==
---
title: "3-report"
author: "bernard-liew"
date: "2020-10-09"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Load package

```{r}
if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse,
               qgraph,
               stats,
               bootnet,
               igraph,
               plotrix,
               mgm,
               furrr,
               cowplot,
               officer,
               flextable,
               arsenal)

# Helper code

meanNsd_transform <- function (x) {

  m <- round (x[[1]][1], 2)
  s <- round (x[[1]][2], 2)

  m_s <- paste0(m, "(", s, ")")

  return (m_s)
}
```

## Report network without group as variable

### ODI labels

```{r}
# ODI custom figure
nodeLabels <- c("Pain Intensity", "Personal Care", "Lifting", "Walking", "Sitting",
                "Standing", "Sleeping", "Social life", "Travelling", "Work/Housework")

nodeLabs <- c("Grp", "Q1.Pain", "Q2.Care", "Q3.Lift", "Q4.Walk", "Q5.Sit",
                "Q6.Std", "Q7.Slp", "Q8.Life", "Q9.Trav", "Q10.Work")

ques <- paste0("Q", 1:10)

node_df <- data.frame("Item" = ques,
                      "Variable" = nodeLabels)
```


## Report network with group as variable

### Load models

Individualised PT coded as 1.

Advise coded as 0.

```{r}
# Model with missing data as input
res <- readRDS("output/mgm_raw.RDS")
# Model with complete imputed data as input
#res <- readRDS("output/com.RDS")

nodes <- colnames (res$data[[1]])
```

## Descriptive statistics 

```{r fig.height=15, fig.width=20, message=FALSE, warning=FALSE}
df <- readRDS("output/dat_odi_nest.RDS")

temp1 <- df$raw %>% select (-data_mgm)
temp2 <- df$com %>% select (-data_mgm) %>% rename (data_com = data)
temp3 <- temp1 %>%
  inner_join(temp2, by = "time")

temp4 <- df$raw %>% select (-data) %>% rename (data = data_mgm)
temp5 <- df$com %>% select (-data) %>% rename (data_com = data_mgm)
temp6 <- temp4 %>%
  inner_join(temp5, by = "time")

df <- list (odi = temp3, 
            odi_mgm = temp6)

rm (list = ls(pattern = "temp"))

df2 <- df$odi_mgm %>%
  select (time, data) %>%
  unnest (data) %>%
  mutate_if (is.factor, as.numeric) %>%
  mutate (grp = factor (grp),
          time = factor (time)) %>%
  pivot_longer(cols = starts_with("Q"),
               names_to = "Items",
               values_to = "val") %>%
  mutate (Items = factor (Items, levels = ques, labels = nodeLabels))

df_plot <- df2 %>%
  group_by(time, grp,Items) %>%
  summarise(Mean = mean (val, na.rm = TRUE),
            Sd = sd (val, na.rm = TRUE))

 f <- ggplot (df_plot) +
  geom_point(aes (x = Items, y = Mean, colour = grp, group = grp),  
             stat = "identity", position=position_dodge(0.5)) +
  geom_errorbar(aes (x = Items, ymin = Mean - Sd, 
                     ymax = Mean + Sd, colour = grp), width = 0, position=position_dodge(0.5)) +
  facet_wrap(~ time, scales = "fixed") +
  scale_color_manual(values = c("red", "blue"), labels = c("IndPT", "Advice")) +  
  labs (x = "ODI Items",
        y = "Mean (SD) score of items",
        colour = "Group") +
  theme_cowplot() + 
  theme(text = element_text(size=16),
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
 
 f

# pdf(width = 15, height = 8, file = "../manuscript_odi_nw/sm_fig1.pdf")
# f
# dev.off()
```

### Plot network

Blue edges - positive correlation

Red edges - negative correlation

The thickness of the edges indicate the magnitude of correlation.

```{r fig.height=15, fig.width=20, message=FALSE, warning=FALSE}

par (mfrow = c(3,2))
p1 <- plot (res$nw[[1]], title = "Baseline", label.cex = 2, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[2]], title = "Week 5", layout = p1$layout, label.cex = 2, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[3]], title = "Week 10", layout = p1$layout, label.cex = 2, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[4]], title = "Week 26", layout = p1$layout, label.cex = 2, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[5]], title = "Week 52", layout = p1$layout, label.cex = 2, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
# plot.new()
# addtable2plot(0,0,node_df, 
#               xpad=1, ypad=1,
#               bty='o',
#               display.rownames = FALSE, 
#               hlines = TRUE,
#               vlines = TRUE)

```

```{r, eval = FALSE, include = FALSE}
pdf(width = 30, height = 20, file = "../manuscript_odi_nw/fig1.pdf")
par (mfrow = c(2,3))
p1 <- plot (res$nw[[1]], title = "Baseline", label.cex = 1, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[2]], title = "Week 5", layout = p1$layout, label.cex = 1, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[3]], title = "Week 10", layout = p1$layout, label.cex = 1, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[4]], title = "Week 26", layout = p1$layout, label.cex = 1, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
plot (res$nw[[5]], title = "Week 52", layout = p1$layout, label.cex = 1, vsize = 15, curve = 0.4, curveAll = TRUE, labels = nodeLabs, title.cex = 4)
# plot.new()
# addtable2plot(0,0,node_df, 
#               xpad=1, ypad=1,
#               bty='o',
#               display.rownames = FALSE, 
#               hlines = TRUE,
#               vlines = TRUE)
dev.off()

```


```{r, include = FALSE, eval = FALSE}
# Easy reporting in manuscript
n = 5

x <- summary (res$edgewts[[n]]) %>%
  ungroup () %>%
  filter (type == "edge" & sample != 0) %>%
  dplyr::select (id, node1, node2, sample, q2.5, q97.5) %>%
  mutate (samp_abs = abs (sample)) %>%
  group_by(node1) %>%
  slice_max (samp_abs, n = 4) %>%
  map_if (is.numeric, round, 2) %>%
  bind_cols()

x_max <- x %>%
  slice_max (sample, n = 1)

nod2 <- x[x$node1 == "grp", ][["node2"]]

for (i in seq_along (nod2)) {
  cat ("Group was related to", x[x$node1 == "grp" & x$node2 == nod2[i], ][["node2"]],
     "with a value of", x[x$node1 == "grp" & x$node2 == nod2[i], ][["sample"]],
     "(95%CI [", x[x$node1 == "grp" & x$node2 == nod2[i],][["q2.5"]], "to", 
     x[x$node1 == "grp" & x$node2 == nod2[i], ][["q97.5"]], "])", "\n")
}


cat ("The edge with the greatest weight magnitude was between", x_max[["node1"]],
     "and",   x_max[["node2"]],
     "with a value of", x_max[["sample"]],
     "([95%CI (", x_max[["q2.5"]], "to", x_max[["q97.5"]], "])")


```

### Plot edge weights stability

#### Confidence interval


```{r fig.height=30, fig.width=20, message=FALSE, warning=FALSE}
w_fig <- map (res$edgewts, plot, order = "sample", CIstyle = "quantiles")

w_fig <- map (w_fig, ~.x + 
                  theme(text = element_text(size = 16)), 
              include = "all", print = FALSE, scale = "relative")

cowplot::plot_grid(plotlist = w_fig, labels = c("Wk0","Wk 5", "Wk 10", "Wk 26", "wk52" ), 
                   vjust = 1, hjust = -1, ncol = 2)



```

```{r, eval = FALSE, include = FALSE}
w_fig <- map (res$edgewts, plot, order = "sample", CIstyle = "quantiles")

w_fig <- map (w_fig, ~.x + 
                  theme(text = element_text(size = 12)), 
              include = "all", print = FALSE, scale = "relative")

pdf(width = 16, height = 25, file = "../manuscript_odi_nw/fig2.pdf")
cowplot::plot_grid(plotlist = w_fig, 
                  labels = c("Baseline","Week 5", "Week 10", "Week 26", "Week52" ), 
                    vjust = 1, hjust = -1, ncol = 2)
dev.off()

```



```{r, eval = FALSE, include = FALSE}
#### Hypothesis testing
wp_fig <- map (res$edgewts, plot, plot = "difference", 
              satistics = "edge", order = "sample", onlyNonZero = TRUE)


wp_fig <- map (wp_fig, ~.x + 
                  theme(text = element_text(size = 16)), 
              include = "all", print = FALSE, scale = "relative")

cowplot::plot_grid(plotlist = wp_fig, labels = c("Wk0","Wk 5", "Wk 10", "Wk 26", "wk52" ), vjust = 1, hjust = -1, ncol = 3)

```


```{r, eval = FALSE, include = FALSE}
wp_fig <- map (res$edgewts, plot, plot = "difference", 
              satistics = "edge", order = "sample", onlyNonZero = TRUE)


wp_fig <- map (wp_fig, ~.x + 
                  theme(text = element_text(size = 14)), 
              include = "all", print = FALSE, scale = "relative")

pdf(width = 12, height = 25, file = "../manuscript_odi_nw/edge_stb_hyp.pdf")
cowplot::plot_grid(plotlist = wp_fig, 
                  labels = c("Baseline","Week 5", "Week 10", "Week 26", "Week52" ), 
                    vjust = 1, hjust = 0, ncol = 2)
dev.off()

```

### Plot centrality

High centrality nodes have strong connections to many other nodes, and act as hubs that connect otherwise disparate nodes to one another. 

Low centrality nodes exist on the periphery of the network, with fewer and weaker connections to other nodes of the
network.

Strength is the sum of the absolute value of its connections with other nodes in the network.

Closeness centrality is defined as the inverse of the sum of the distances of the focal node from all the other nodes in the network. Closeness is the average shortest path between a given node and the remaining nodes in the network. Nodes with higher closeness are more proximally connected to the rest of the network.

Betweenness is the number of times in which a given node lies on the shortest path between two other nodes.

The greater the value of centrality indices to one, the more important the variable.

```{r fig.height=20, fig.width=15, message=FALSE, warning=FALSE}
# Plot centrality
c_fig <- map (res$nw, centralityPlot, include = c("Closeness", "Strength", "Betweenness"),  
              print = FALSE, scale = "relative", labels = nodeLabs) %>%
  map (~.x + 
                  scale_x_continuous(breaks= c(0, 0.5, 1), lim = c(0, 1)) + 
                  theme(text = element_text(size = 20), 
                        axis.text.x = element_text(angle = 45, hjust = 1)), 
              include = "all", print = FALSE, scale = "relative")

#tiff(width = 15, height = 15, units = "in", res = 100, file = "output/odi_strength.tiff")
cowplot::plot_grid(plotlist = c_fig, labels = c("Wk0","Wk 5", "Wk 10", "Wk 26", "wk52" ), vjust = 1, hjust = 0, ncol = 2)
#dev.off()

```

```{r eval = FALSE, include = FALSE}
c_fig <- map (res$nw, centralityPlot, include = c("Closeness", "Strength", "Betweenness"), 
              print = FALSE, scale = "relative", labels = nodeLabs) %>%
  map (~.x + 
                  scale_x_continuous(breaks= c(0, 0.5, 1), lim = c(0, 1)) + 
                  theme(text = element_text(size = 20), 
                        axis.text.x = element_text(angle = 45, hjust = 1)), 
              include = "all", print = FALSE, scale = "relative")

pdf(width = 20, height = 15, file = "../manuscript_odi_nw/fig3.pdf")
cowplot::plot_grid(plotlist = c_fig, 
                  labels = c("Baseline","Week 5", "Week 10", "Week 26", "Week52" ), 
                    vjust = 1, hjust = 0, ncol = 3)
dev.off()

```

```{r, eval = FALSE, include = FALSE}

n_time <- c("baseline", "week 5", "week 10", "week 26", "week 52")

for (n in seq_along (n_time)) {
  
  c_val <- c_fig[[n]]$data %>%
    group_by(node) %>%
    mutate (ave = mean (value)) %>%
    ungroup () %>%
    slice_max (ave, n = 6) %>%
    pull (node) %>%
    unique () %>%
    as.character()
  
  cat ("At", n_time[n], "across all three centrality measures, the two nodes with the greatest value was",     c_val[[1]], "and", c_val[[2]], "\n")

}


```


### Plot centrality stability

Is the centrality order stable?

```{r fig.height=20, fig.width=15, message=FALSE, warning=FALSE}
# Plot centrality stability
s_fig <- map (res$centr_stb, plot, statistics = c("closeness", "strength", "betweenness"))

s_fig <- map (s_fig, ~.x + 
                ylab ("Ave Corr") + 
                  theme(text = element_text(size = 20), 
                        axis.text.x = element_text(angle = 90, hjust = 1)), 
              include = "all", print = FALSE, scale = "relative")


#tiff(width = 15, height = 15, units = "in", res = 100, file = "output/odi_stability.tiff")
cowplot::plot_grid(plotlist = s_fig, labels = c("Wk0","Wk 5", "Wk 10", "Wk 26", "wk52" ), vjust = 1, hjust = -1, ncol = 2)
#dev.off()

```

```{r eval = FALSE, include = FALSE}
# Plot centrality stability
s_fig <- map (res$centr_stb, plot, statistics = c("closeness", "strength", "betweenness"))

s_fig <- map (s_fig, ~.x + 
                ylab ("Ave Corr") + 
                  theme(text = element_text(size = 20), 
                        axis.text.x = element_text(angle = 90, hjust = 1)), 
              include = "all", print = FALSE, scale = "relative")

pdf(width = 20, height = 15, file = "../manuscript_odi_nw/fig4.pdf")
cowplot::plot_grid(plotlist = s_fig, 
                  labels = c("Baseline","Week 5", "Week 10", "Week 26", "Week52" ), 
                    vjust = 1, hjust = 0, ncol = 3)
dev.off()

```

### Get CS coefficient

The stability of centrality estimation, and results in a centrality-stability coefficient (CS-coefficient) that should not be lower than 0.25 and preferably above 0.5

```{r, results = "asis", message=FALSE, warning=FALSE}
cs_coef <- res %>%
  select (time, cor_stb) %>%
  unnest () %>%
  ungroup() %>%
  mutate (measure = rep (c("betweenness", "closeness", "edge", "expectedInfluence", "strength"), 5)) %>%
  mutate (CS = round (cor_stb, 2)) %>%
  filter (measure %in% c("betweenness", "closeness","strength")) %>%
  dplyr::select (time, measure, CS)

cs_coef %>%
  knitr::kable (caption = "Stability of centrality indices")

```

```{r, eval = FALSE, include = FALSE}
cs_coef %>%
    group_by(time) %>%
    slice_max (CS, n = 1)
  
```

```{r eval = FALSE, include = FALSE}
my_path <- paste0("../manuscript_odi_nw/table_", 
                  "CScoef",
                  ".docx")


ft <- flextable(cs_coef) %>%
  set_caption(caption = "Table 1. CS Coefficient") %>%
  autofit()

my_doc <- read_docx()  %>% 
  body_add_flextable(ft) %>%
  body_end_section_landscape()

print (my_doc, target = my_path)
```
